Fix #319

+++ scripts/deploy.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 #
 # deploy.sh - Application deployment script
 # Pulls latest code, builds, and deploys to production
@@ -29,7 +30,8 @@ check_dependencies
 
 # Clean previous deployment artifacts
 log_message "Cleaning previous build artifacts..."
-rm -rf ${DEPLOY_DIR}/dist
+[ -z "$DEPLOY_DIR" ] && exit 1
+rm -rf "${DEPLOY_DIR}/dist"
 rm -rf ${DEPLOY_DIR}/node_modules/.cache
 
 # Pull latest code
